pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "l2r"
version = "0.1.0"
description = "Learning to Reduce: A hierarchical neural framework to address the scalability limitations of search space reduction in vehicle routing problems"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "L2R Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "numpy>=1.23.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.0.262",
]

[tool.setuptools]
packages = ["l2r"]

[tool.ruff]
select = ["E", "F", "I", "N", "W", "D"]
ignore = ["D203", "D213"]
line-length = 100

[tool.ruff.isort]
known-first-party = ["l2r"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.black]
line-length = 100
target-version = ["py38"] 